requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-generativeai==0.3.1
python-dotenv==1.0.0
markdown==3.5
google-genai
Pillow==10.1.0